[Core] Support anonymous parameter types
Cucubumber expressions installs the  anonymous parameter type into the registry:
```java
ParameterType(
  name: ""
  regexps: [\.*\]
  transformer: s -> s
  useForSnippets: false
  preferForRegexMatch: true
)
```

This allows the step definition to hint the type to which the captured text
should be converted. When used in conjunction with an object mapper
this makes it easier to quickly define step definitions without the need
to define parameter types for each type. For example the following
will work without the need to configure additional parameter types.

```java
Given("^there is some date (.*)$", (Date a) -> { })

Given("there is some date {}", (Date a) -> { })

Given("there is some transaction {} {}", (BigDecimal a, Currency b) -> { })

Given("there is some type {}", (TypeEnum a) -> { })
```

To make the dynamic transform possible type hints are passed to the cucumber
expression when matching a step. Using these hints the `s -> s` transform is
replaced with `s -> defaultTransform.transform(s, typeHint)`. A simple
implementation of the `defaultTransform` is provided that supports a selection
of primitives but should generally speaking be replaced with a proper object mapper.

Fixes: https://github.com/cucumber/cucumber-jvm/issues/1478